+++ src/eulertools/exceptions.py
@@ -1,0 +1,14 @@
+class MissingProjectRootError(FileNotFoundError):
+    pass
+
+
+class FailedRunError(RuntimeError):
+    pass
+
+
+class InvalidLanguageError(ValueError):
+    pass
+
+
+class InvalidProblemError(ValueError):
+    pass